+++ CHANGELOG.md
@@ -7,6 +7,7 @@ MumbleDJ Changelog
 * Restructured project for easier imports.
 * Added .gitignore for pymumble.
 * Now successfully connects to Mumble servers.
+* Added command parsing.
 
 ### September 11, 2014
 * mumble-music-bot repository created.